# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie Auditions API",
    "description": "API Documentation for Movie Auditions Backend",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Casting Director"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "manage_movies"
            }
          },
          "description": {
            "type": "string",
            "example": "Can manage movies and auditions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-12-04T12:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-12-04T12:00:00Z"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/admin/roles": {
      "get": {
        "tags": ["Admin Roles"],
        "summary": "Get list of roles (Admin)",
        "description": "Returns list of all roles for admin",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Role"
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Roles retrieved successfully."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": ["Admin Roles"],
        "summary": "Store a new role (Admin)",
        "description": "Creates a new role",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "permissions"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Casting Director"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "manage_movies"
                    }
                  },
                  "description": {
                    "type": "string",
                    "example": "Can manage movies and auditions"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Role"
                    },
                    "message": {
                      "type": "string",
                      "example": "Role created successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/admin/roles/{id}": {
      "get": {
        "tags": ["Admin Roles"],
        "summary": "Get role information (Admin)",
        "description": "Returns role data",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Role id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Role"
                    },
                    "message": {
                      "type": "string",
                      "example": "Role retrieved successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Role not found"
          }
        }
      },
      "put": {
        "tags": ["Admin Roles"],
        "summary": "Update existing role (Admin)",
        "description": "Updates a role",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Role id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "permissions"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Casting Director"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "manage_movies"
                    }
                  },
                  "description": {
                    "type": "string",
                    "example": "Can manage movies and auditions"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Role"
                    },
                    "message": {
                      "type": "string",
                      "example": "Role updated successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Role not found"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      },
      "delete": {
        "tags": ["Admin Roles"],
        "summary": "Delete role (Admin)",
        "description": "Deletes a role",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Role id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Role deleted successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Role not found"
          }
        }
      }
    },
...
  }
}